Add audio/video section for creation separate from consumption
There are LOT of options in this category. I've only picked a few
that I have personal experience with and which I think fit the
spirit of this project. I understand opinions may differ for the
slightly more well-known companies here, like Blackmagic. None of
these are precious to me except maybe the open source ones since
I think they deserve the attention.